backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9,<1
requests>=2.32,<3
beautifulsoup4>=4.12,<5
python-docx>=1.1,<2
python-pptx>=0.6.23,<1
pypdf>=5,<6
bibtexparser>=1.4,<2
pytest>=8,<9
httpx>=0.28,<1
